Refactor: user, auth 관련 로직을 domain 폴더로 이동
