+++ server/services/auth/src/domain/auth/auth.service.ts
@@ -3,7 +3,7 @@ import { HttpStatus, Injectable } from '@nestjs/common';
 import { User } from '@prisma/client';
 import { firstValueFrom, map } from 'rxjs';
 
-import { UserService } from 'src/user/user.service';
+import { UserService } from 'src/domain/user/user.service';
 import RequestLoginDto from './dto/request-login.dto';
 import { JwtService } from './jwt.service';
