+++ server/services/auth/src/domain/user/user.service.ts
